src/open_ticket_ai/pyproject.toml
[project]
name = "open-ticket-ai"
version = "1.0.0rc1"
description = "AI enhancements for open source ticket systems"
readme = "README.md"
requires-python = ">=3.12"

authors = [
    { name = "Tobias Bück", email = "tab@softoft.de" },
]
license = { text = "LGPL-2.1-only" }

classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU Lesser General Public License v2 (LGPLv2)",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "pyyaml>=6.0.2",
    "pydantic~=2.11.7",
    "pydantic_core",
    "typing_extensions",
    "pytest~=8.4.0",
    "pytest-asyncio~=0.24.0",
    "pluggy",
    "pygments",
    "injector~=0.22.0",
    "requests~=2.32.4",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "openai>=1.108.1",
    "mdxlate>=0.1.0",
    "pathspec>=0.12.1",
    "docstring-parser>=0.17.0",
    "pyfiglet>=1.0.4",
    "tenacity>=9.0.0",
    "prefect~=3.4.20",
    "typer>=0.9.0",
    "apscheduler>=3.10.0",
]

[project.urls]
Homepage = "https://open-ticket-ai.com"
Repository = "https://github.com/Softoft-Orga/open-ticket-ai"

[project.optional-dependencies]
dev = [
    "ruff",
    "mypy",
    "pytest~=8.4.0",
]

[project.scripts]
otai = "open_ticket_ai.cli.main:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["open_ticket_ai"]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "PL", "PERF", "TRY"]
ignore = ["TRY003"]
fixable = ["ALL"]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["PLR2004"]

[tool.mypy]
python_version = "3.13"
strict = true
plugins = ['pydantic.mypy']

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
markers = [
    "unit: schnelle, isolierte Tests",
    "integration: core + plugin zusammen",
    "contract: plugin-api-vertrag",
    "e2e: end-to-end flows",
    "slow: dauert länger"
]
